Test cases for Adult and Minor classes finished; Minor changes to each class to make it work better; Removed an infinite loop where the player contained the product and the product contained the player so toString() and equals() wouldn't work

+++ src/main/java/org/example/Customer.java
@@ -14,7 +14,7 @@ public abstract class Customer {
         this(name, gender, 0);
     }
 
-    public Customer(String name, Gender gender, int money) {
+    public Customer(String name, Gender gender, double money) {
         this.name = name;
         this.gender = gender;
         this.money = money;
